+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.8.0] - Not released
+### Changed
+- Invitations from inactive (i.e. in some 'gone' status) users stop working.
+
 ### Fixed
 - Invitations are no longer deleted when their author's data is deleted. These
   invitations are instead anonymized. This is important for keeping a connection